refactor: code modularization
