+++ src/app/app.ts
@@ -2,47 +2,13 @@ import "../public/resources/css/app.css";
 import "../public/resources/css/index.css";
 
 import loadIcons from "./helpers/fontAwesomeIcons";
-import getHTMLElement from "./helpers/getHTMLElement";
-import StoreApi from "./services/StoreApi";
-import renderSuggestions from "./views/renderSuggestions";
+import loadComponents from "./helpers/loadComponents";
 import loadProductDetails from "./controllers/product-details";
 import loadProductList from "./controllers/product-list";
 import loadSignupForm from "./controllers/signup";
-import Header from "./components/Header";
-import Brand from "./components/Brand";
-import Footer from "./components/Footer";
 
 loadIcons();
-
-customElements.define("header-component", Header);
-customElements.define("brand-component", Brand);
-customElements.define("footer-component", Footer);
-
-const header: HTMLElement = getHTMLElement("header");
-const navigation: HTMLElement = getHTMLElement(".navigation", header);
-const suggestions: HTMLUListElement = getHTMLElement(
-  "#suggestions",
-  navigation,
-) as HTMLUListElement;
-
-header?.querySelector(".burger-menu")?.addEventListener("click", () => {
-  navigation?.classList.toggle("hidden");
-});
-
-const searchBar: HTMLInputElement = getHTMLElement("#search-bar", navigation) as HTMLInputElement;
-searchBar.addEventListener("input", async function () {
-  const value = this.value.trim();
-
-  if (value == "") {
-    suggestions.innerHTML = "";
-    return;
-  }
-
-  const res = await StoreApi.filterCatalog(value);
-  const data = await res.json();
-
-  renderSuggestions(data, suggestions);
-});
+loadComponents();
 
 switch (window.location.pathname) {
   case "/product-list":
